Fix [#293] 수정 뷰 수정
