--- YELLO-iOS/YELLO-iOS/Presentation/Profile/Main/View/EditProfileView.swift
@@ -1,34 +1,0 @@
-//
-//  EditProfileView.swift
-//  YELLO-iOS
-//
-//  Created by 지희의 MAC on 1/25/24.
-//
-
-import UIKit
-
-import SnapKit
-import Then
-
-class EditProfileView: UIView {
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        setUI()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    private func setUI(){
-        setStlye()
-        setLayout()
-    }
-    
-    private func setStlye() {
-    }
-    
-    private func setLayout() {
-        <#code#>
-    }
-}
